pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes"
version = "0.1.0"
description = "Multi-profile Binance trading bot (Sentinel / Equilibrium / Vortex)"
requires-python = ">=3.10"

dependencies = [
  "python-binance>=1.0.19",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "python-telegram-bot>=20.7",
  "python-telegram-bot[job-queue]>=20.7",
  "sqlalchemy>=2.0",
  "pymysql>=1.1.0"
]

# =========================
# CLI ENTRYPOINTS
# =========================
[project.scripts]
hermes = "hermes.cli:main"

# =========================
# SETUPTOOLS CONFIG
# =========================
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["hermes", "hermes.*"]
